src/backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt --python-version 3.10
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.14
    # via fastapi-app (pyproject.toml)
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
async-timeout==5.0.1
    # via
    #   aiohttp
    #   asyncpg
asyncpg==0.30.0
    # via fastapi-app (pyproject.toml)
attrs==25.3.0
    # via aiohttp
azure-core==1.35.0
    # via azure-identity
azure-identity==1.23.1
    # via fastapi-app (pyproject.toml)
backports-datetime-fromisoformat==2.0.3
    # via marshmallow
certifi==2025.7.14
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
contourpy==1.3.2
    # via matplotlib
cryptography==45.0.5
    # via
    #   azure-identity
    #   msal
    #   pyjwt
cycler==0.12.1
    # via matplotlib
environs==14.2.0
    # via fastapi-app (pyproject.toml)
exceptiongroup==1.3.0
    # via anyio
fastapi==0.116.1
    # via fastapi-app (pyproject.toml)
fonttools==4.60.1
    # via matplotlib
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
geoalchemy2==0.18.0
    # via fastapi-app (pyproject.toml)
greenlet==3.2.3
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
    #   yarl
kiwisolver==1.4.9
    # via matplotlib
marshmallow==4.0.0
    # via environs
matplotlib==3.10.7
    # via fastapi-app (pyproject.toml)
msal==1.32.3
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
numpy==2.2.6
    # via
    #   contourpy
    #   matplotlib
packaging==25.0
    # via
    #   geoalchemy2
    #   matplotlib
pillow==11.3.0
    # via matplotlib
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pyjwt==2.10.1
    # via msal
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
python-dotenv==1.1.1
    # via
    #   fastapi-app (pyproject.toml)
    #   environs
requests==2.32.4
    # via
    #   azure-core
    #   msal
six==1.17.0
    # via
    #   azure-core
    #   python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
    # via
    #   fastapi-app (pyproject.toml)
    #   geoalchemy2
starlette==0.47.2
    # via fastapi
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   anyio
    #   azure-core
    #   azure-identity
    #   environs
    #   exceptiongroup
    #   fastapi
    #   marshmallow
    #   multidict
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via fastapi-app (pyproject.toml)
yarl==1.20.1
    # via aiohttp
